Organize CD-ROM tests to mirror source directories
